Add ability to extract the Expression that has been built for re-use
This enables a user to build and save an expression for later use.

Two methods have been added to the Api, these are `AsExpression()` and
`AsFunc()`.  The reason for this is because, depending on where you wish
to apply the expression (on `IEnumerable` or `IQueryable`) changes the
form in which you need the expression.  Granted a user could simply
`.Compile()` the expression to get the func